Struct list_metrics::Client

pub(crate) struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> {
    handle: Arc<Handle<C, M, R>>,
}

An ergonomic service client for GraniteServiceVersion20100801.

This client allows ergonomic access to a GraniteServiceVersion20100801-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Fields
handle: Arc<Handle<C, M, R>>
Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Client<C, M, R>

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn delete_alarms(&self) -> DeleteAlarms<C, M, R>

Constructs a fluent builder for the DeleteAlarms operation.

See DeleteAlarms for more information about the operation and its arguments.

pub fn delete_anomaly_detector(&self) -> DeleteAnomalyDetector<C, M, R>

Constructs a fluent builder for the DeleteAnomalyDetector operation.

See DeleteAnomalyDetector for more information about the operation and its arguments.

pub fn delete_dashboards(&self) -> DeleteDashboards<C, M, R>

Constructs a fluent builder for the DeleteDashboards operation.

See DeleteDashboards for more information about the operation and its arguments.

pub fn delete_insight_rules(&self) -> DeleteInsightRules<C, M, R>

Constructs a fluent builder for the DeleteInsightRules operation.

See DeleteInsightRules for more information about the operation and its arguments.

pub fn delete_metric_stream(&self) -> DeleteMetricStream<C, M, R>

Constructs a fluent builder for the DeleteMetricStream operation.

See DeleteMetricStream for more information about the operation and its arguments.

pub fn describe_alarm_history(&self) -> DescribeAlarmHistory<C, M, R>

Constructs a fluent builder for the DescribeAlarmHistory operation.

See DescribeAlarmHistory for more information about the operation and its arguments.

pub fn describe_alarms(&self) -> DescribeAlarms<C, M, R>

Constructs a fluent builder for the DescribeAlarms operation.

See DescribeAlarms for more information about the operation and its arguments.

pub fn describe_alarms_for_metric(&self) -> DescribeAlarmsForMetric<C, M, R>

Constructs a fluent builder for the DescribeAlarmsForMetric operation.

See DescribeAlarmsForMetric for more information about the operation and its arguments.

pub fn describe_anomaly_detectors(&self) -> DescribeAnomalyDetectors<C, M, R>

Constructs a fluent builder for the DescribeAnomalyDetectors operation.

See DescribeAnomalyDetectors for more information about the operation and its arguments.

pub fn describe_insight_rules(&self) -> DescribeInsightRules<C, M, R>

Constructs a fluent builder for the DescribeInsightRules operation.

See DescribeInsightRules for more information about the operation and its arguments.

pub fn disable_alarm_actions(&self) -> DisableAlarmActions<C, M, R>

Constructs a fluent builder for the DisableAlarmActions operation.

See DisableAlarmActions for more information about the operation and its arguments.

pub fn disable_insight_rules(&self) -> DisableInsightRules<C, M, R>

Constructs a fluent builder for the DisableInsightRules operation.

See DisableInsightRules for more information about the operation and its arguments.

pub fn enable_alarm_actions(&self) -> EnableAlarmActions<C, M, R>

Constructs a fluent builder for the EnableAlarmActions operation.

See EnableAlarmActions for more information about the operation and its arguments.

pub fn enable_insight_rules(&self) -> EnableInsightRules<C, M, R>

Constructs a fluent builder for the EnableInsightRules operation.

See EnableInsightRules for more information about the operation and its arguments.

pub fn get_dashboard(&self) -> GetDashboard<C, M, R>

Constructs a fluent builder for the GetDashboard operation.

See GetDashboard for more information about the operation and its arguments.

pub fn get_insight_rule_report(&self) -> GetInsightRuleReport<C, M, R>

Constructs a fluent builder for the GetInsightRuleReport operation.

See GetInsightRuleReport for more information about the operation and its arguments.

pub fn get_metric_data(&self) -> GetMetricData<C, M, R>

Constructs a fluent builder for the GetMetricData operation.

See GetMetricData for more information about the operation and its arguments.

pub fn get_metric_statistics(&self) -> GetMetricStatistics<C, M, R>

Constructs a fluent builder for the GetMetricStatistics operation.

See GetMetricStatistics for more information about the operation and its arguments.

pub fn get_metric_stream(&self) -> GetMetricStream<C, M, R>

Constructs a fluent builder for the GetMetricStream operation.

See GetMetricStream for more information about the operation and its arguments.

pub fn get_metric_widget_image(&self) -> GetMetricWidgetImage<C, M, R>

Constructs a fluent builder for the GetMetricWidgetImage operation.

See GetMetricWidgetImage for more information about the operation and its arguments.

pub fn list_dashboards(&self) -> ListDashboards<C, M, R>

Constructs a fluent builder for the ListDashboards operation.

See ListDashboards for more information about the operation and its arguments.

pub fn list_metrics(&self) -> ListMetrics<C, M, R>

Constructs a fluent builder for the ListMetrics operation.

See ListMetrics for more information about the operation and its arguments.

pub fn list_metric_streams(&self) -> ListMetricStreams<C, M, R>

Constructs a fluent builder for the ListMetricStreams operation.

See ListMetricStreams for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn put_anomaly_detector(&self) -> PutAnomalyDetector<C, M, R>

Constructs a fluent builder for the PutAnomalyDetector operation.

See PutAnomalyDetector for more information about the operation and its arguments.

pub fn put_composite_alarm(&self) -> PutCompositeAlarm<C, M, R>

Constructs a fluent builder for the PutCompositeAlarm operation.

See PutCompositeAlarm for more information about the operation and its arguments.

pub fn put_dashboard(&self) -> PutDashboard<C, M, R>

Constructs a fluent builder for the PutDashboard operation.

See PutDashboard for more information about the operation and its arguments.

pub fn put_insight_rule(&self) -> PutInsightRule<C, M, R>

Constructs a fluent builder for the PutInsightRule operation.

See PutInsightRule for more information about the operation and its arguments.

pub fn put_metric_alarm(&self) -> PutMetricAlarm<C, M, R>

Constructs a fluent builder for the PutMetricAlarm operation.

See PutMetricAlarm for more information about the operation and its arguments.

pub fn put_metric_data(&self) -> PutMetricData<C, M, R>

Constructs a fluent builder for the PutMetricData operation.

See PutMetricData for more information about the operation and its arguments.

pub fn put_metric_stream(&self) -> PutMetricStream<C, M, R>

Constructs a fluent builder for the PutMetricStream operation.

See PutMetricStream for more information about the operation and its arguments.

pub fn set_alarm_state(&self) -> SetAlarmState<C, M, R>

Constructs a fluent builder for the SetAlarmState operation.

See SetAlarmState for more information about the operation and its arguments.

pub fn start_metric_streams(&self) -> StartMetricStreams<C, M, R>

Constructs a fluent builder for the StartMetricStreams operation.

See StartMetricStreams for more information about the operation and its arguments.

pub fn stop_metric_streams(&self) -> StopMetricStreams<C, M, R>

Constructs a fluent builder for the StopMetricStreams operation.

See StopMetricStreams for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(
conf: Config,
conn: C
) -> Client<C, AwsMiddleware, Standard>

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Client<DynConnector, AwsMiddleware, Standard>

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Client<DynConnector, AwsMiddleware, Standard>

Creates a new client from the service Config.